rename metrics package
