+++ modules/metrics/src/test/java/org/springside/modules/metrics/ReporterTest.java
@@ -11,11 +11,11 @@ import java.util.concurrent.TimeUnit;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.springside.modules.metrics.Timer.TimerContext;
-import org.springside.modules.metrics.report.ConsoleReporter;
-import org.springside.modules.metrics.report.GraphiteReporter;
-import org.springside.modules.metrics.report.ReportScheduler;
-import org.springside.modules.metrics.report.Reporter;
-import org.springside.modules.metrics.report.Slf4jReporter;
+import org.springside.modules.metrics.reporter.ConsoleReporter;
+import org.springside.modules.metrics.reporter.GraphiteReporter;
+import org.springside.modules.metrics.reporter.ReportScheduler;
+import org.springside.modules.metrics.reporter.Reporter;
+import org.springside.modules.metrics.reporter.Slf4jReporter;
 import org.springside.modules.metrics.utils.Clock.MockClock;
 
 public class ReporterTest {
